mm-video: vidc: Possibly send not-coded-VOPs to the driver
Previously we were dropping the not coded VOPs in the component because
the legacy hardware didn't support decoding it.  Since the hardware now
supports decoding N-C-VOPs, send it to the hardware if the client
insists on it being decoded.  A side effect of this is FBDs which have
duplicate YUVs and timestamps.  The client is expected deal with this
side effect.

CRs-Fixed: 714038
Change-Id: I44aa4f1f26568cf59cc76751802b8aa2373fcb69