mm-video-v4l2: vdec: Request buffer for input port after client reconfigures
When OMX client reconfigures, buffer requirement should be called
again, else ETB will fail as the buffers are not prepared.

Release driver buffers when all component buffers are released.
This is needed as component may transition back to idle state
from loaded state, and prepare_buf will fail if the buffers
are not unregistered from driver.

Change-Id: I02762574277a36f7ba334843b91cb77953cad5c9